Update patch set 1
No score.

(Sorry it took so long to respond to this. I have been trying to respond
for the last 10 hours, and kept getting 500 errors from Gerrit. I
finally figured out it was because I needed to respond to the actual
comment inline to the code rather than to the "No score." comment on the
main page.)

The reason why I didn't do any error checking in this block of code is
because error checking in this entire file is random: 

there is no basis of comparison to know "this is the standard error
checking concept". :( Just scanning around the file, the 

most number of commands are in the category of doing no argument
checking and returning 0 even in case of error.

I have now done an analysis of all commands and how they respond to
errors, and I agree: "chroot" and "chdir" feel more like 

the (very, very few) commands that return -errno for most errors.
Whether it should segment fault with incorrect arguments or 

return an error is still in question, but the (statistically
insignificant data) wants to pull me towards "segment fault". ;P

Unfortunately, there are no instructions in the repo documentation for
how I'd amend this patch other than to delete it and 

put an entirely new one up, which seems wrong (as then we lose things
like comment history)... is that what I'm supposed to 

do?

Regardless, in case anyone is interested, here is my analysis. Note
that, unless otherwise marked, commands segment fault 

given some number of arguments (one command: mkdir, looks like it
checks, but doesn't actually):

return 0 for most errors:
   class_start
   class_stop
   device (returns -1 for bad arguments)
   export
   restart
   setprop
   start
   stop

return -1 for most errors:
   domainname
   hostname
   ifup
   loglevel (returns -1 for bad arguments)
   mount
   setrlimit
   symlink
   write

return -errno for most errors:
   chmod
   chown (returns -1 for bad arguments)
   mkdir

aren't even implemented, and what they return:
   exec: returns -1
   import: returns -1
   trigger: returns 0...


On 2008/11/19 20:16:07, "Alexey Tarasov" <tarasov wrote:
> Just curious, why do you always return zero?

Patch-set: 1