Update patch set 1
No score.

While I agree with these comments, they are in code that is not affected
by this patch, and in a perfect world (where repo/Gerrit made this
simple), would have been submitted as a separate changeset rather than
as comments. If this change goes through I will happily fix this other
issue as well, but changing those lines will conflict in the context of
this patch and I have not seen any documentation on how to create
dependent changes with repo/Gerrit (even though I know it is possible as
I've seen other developers do it).

Patch-set: 1