Update patch set 1
I would prefer that you didn't submit this.

I'm concerned that making Android explicitly rely on binaries that
aren't part of the system goes against the notion that the system should
be self-contained.

I'm even more concerned that this relies on a hard-coded path to a
location on /system, which is a read-only partition. That would prevent
users from installing the binary in question, which would make this
patch essentially useless on its own... and pointing to a location that
users can write to would open a huge root hole since we'd now be
running as root a user-supplied binary.

I'd prefer to see busybox go into the system first if it makes sense (or
a replacement just for the bit of functionality that's necessary here),
and then the code that uses it.

Patch-set: 1