Update patch set 1
I would prefer that you didn't submit this.

The decision about /system/bin/iptables isn't quite mine to make. In a
way, the best way to know is to discuss this on the android-platform
mailing list. The second best way to know is to submit a patch that adds
it to production device images, with an explanation about what it's
needed for and why the added functionality justifies the expense of
storage space, and see whether that patch gets accepted.

In the big picture of what this change here (3123) is trying to achieve,
trying to go the iptables route (no pun intended) seems to be your most
likely path to success.

As for the SDK, I have no idea.

Patch-set: 1