Update patch set 1
No score.

(Lets see if 'reply' works yet.. Nope. 500 error. Way to go..)

On 2008/12/03 00:49:29, "Brian Swetland" <swetland wrote:
> - depends on external entities not part of the core platform

> - would be better supported as part of a board-specific init.rc

Its not board specific, why would it be in a board-specific init? pand
is a generic daemon that is part of bluez and works with any bt
implementation that uses bluez. (And it is built as part of the platform
build. Ditto iptables. I am planning on modifying the iptables line to
/system/bin/iptables when I do rev 2, once I get time to has out the
speed problems.)

Udhcpd can be left out if you want to continue reinventing the wheel by
writing a dhcp daemon (or, alternately, if you want to force everyone to
use static IP for tethering - let me know, thats a moderately smaller
patch I can submit in a few seconds.) But nobody has presented any
technical or legal reasons to do that, so..

Patch-set: 1