Update patch set 2
Patch Set 2: No score

@Jean-Baptiste: thanks!

> Make allows for a space after the variable name, and that's
> our common style so it'd be great to continue using it.

I was continuing the style of the existing Android.mk for
sh, and the style of mksh’s Makefile for mksh/Android.mk.
Otherwise I would need to change all of it, which intro-
duces a lot of unnecessary noise.

I would *greatly* prefer if (you? me? someone?) changed
the style, if needed, afterwards in a separate commit.

Is that ok for you?

> Line 63: # only if TARGET_SHELL=ash: symlink /system/bin/sh → ash
> Is this arrow a non-ASCII character? Looks pretty, but
> I'd prefer ASCII-art

OK, will remember that for my next upload. It’s UTF-8, yes.

Patch-set: 2
Label: Code-Review=0