Update patch set 1
Patch Set 1: (10 inline comments)

Instead of adding a new function that essentially duplicates the functionality of ifc_add_host_route, but supports IPv6 as well, could you instead add a more generic add_route function that would support both host routes and network routes, and both IPv4 and IPv6? That would seem to be more useful in the general case, and is not much harder to implement. Also, don't take the address type as a parameter - it's implicit from the string format. So you could have something like this:

extern int ifc_add_route(const char *ifname, const char *addr, const char *prefixlen);

This function would use getnameinfo() with the NI_NUMERICHOST flag to extract the address family from the string, and then call the appropriate function to add IPv4 or IPv6 routes as appropriate. 

That way you wouldn't have to define the IPv4 and IPv6 constants anywhere, either in the C code or in the Java code (which is good, since they're not typically defined in other code - C uses AF_INET / AF_INET6 and Java uses classes). You'd also not have to worry about having to deal with handling inet_pton failures for invalid address formats: if the address is invalid, getnameinfo will fail right away with the proper error and you can pass that back to the caller.

Also, instead of having host routes and default routes, why not just have one function that creates an IPv6 route, and one that creates an IPv4 route? If the prefix length is 32 for IPv4 or 16 for IPv6 the function would create a host route, otherwise a gatewayed route. That would have the advantage that you would only need to call if_nametoindex in one place.

Please use "family" instead of "address type" to indicate IPv4 vs. IPv6 - that is the convention used by most networking code.

Finally, please do check for errors and set errno if appropriate. That way you can make higher-level changes that can pass the right error code to JNI which can at least throw a generic exception with a meaningful error message  (which you can obtain as sys_errlist[errno]).

Comments on the code are below but if you simplify the code as I suggest you probably won't have to do many of them.

Patch-set: 1