Update patch set 1
Patch Set 1:

Yes, it's gonna be bitter either way: no matter which option we pick (make it dependent on linux or on requiring it), x86 builds will be broken on MacOS, and that's gonna remain true until we either remove the dependency on libdiskconfig or put one together that doesn't require linux-specific host headers.

I do prefer to condition it on linux, as it means that breakages in that area will surface even if we don't build x86 targets.

Patch-set: 1