Update patch set 6
Patch Set 6: Verified

Verifies fine.

FWIW, the tentative optimization might not have optimize that much. On ARM, function calls are cheap. Also, it was reading memory out of order. Finally, testing for = boils down to testing on the length of the property, which might not be as good a factor for early differentiation as the first few characters of the property itself.

Patch-set: 6
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=+1