Update patch set 7
Patch Set 7: Looks good to me, but someone else must approve

A very minor issue: could I suggest to reset the date of the commit to reflect the real time it was made? It's "Date:   Thu Jun 3 19:53:01 2010 +0200" now. But the log said it's from 2010/07/15. A little weird.

Patch-set: 7
Label: Code-Review=+1