Update patch set 7
Patch Set 7:

I was going to suggest David (due to the goldfish shellscript), but it seems he already +1'd it.

It's an odd mix of a shellscript change, a route.c change (which the shellscript doesn't depend on), and a whitespace change in a makefile.

I have no objection to any of the three changes here (though I do scratch my head at why they're bundled), and am happy to +2 it if you just need somebody to do so.

Patch-set: 7
Reviewer: Brian Swetland <1001286@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0