Update patch set 4
Patch Set 4: Fails

Indeed, there's an issue with the copyright/license header. We very strongly prefer to have the same copyright notice on all files ("The Android Open Source Project"). Using different notices can make other companies nervous about using the files in question.

Sadly, looking around it appears that we overlooked a similar issue with col32cb16blend.S and its neon variant. We'd very much appreciate if ARM could quickly upload a patch that changes the copyright notice on those files to "The Android Open Source Project" as well so that we don't find ourselves in the uncomfortable situation where we need to discuss whether to keep the files in question or delete them.

Sorry about overlooking this issue, and thanks for understanding.

Patch-set: 4