Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(6 inline comments)

Apart from the obvious and trivial (indentation, spelling, naming convention), I am not really sure I like the idea of this.

Please correct me if I misunderstood something. This patch will simply create one thread for init.rc and another one for init.HARDWARE.rc, and execute the commands in parallel? Will that result in a large gain, and not just increase complexity? While some operations may take long time (such as mount, insmod perhaps, or exec foo), most of them are really quick.

Also, there is a lot to gain of having a sequential "easy to understand" init sequence. It's robust to start with.

Also, the 'exec' command (while not implemented in the stock Android, but there is a contribution uploaded for review) loses its meaning since it was supposed to run "synchronously" (and "alone"). It can be used for doing FOTA upgrades during bootup, or anything else where you want to halt the rest of the execution.

Just my 2 cents.

Patch-set: 1