Update patch set 5
Patch Set 5:

This change is going to conflict with our internal versions of this file.  I can check this change in so it's available in open source, but this change will *not* be in future Android releases.  I'll need to manually modify our non-open-source files to use uevent_checked_recv().

Vernon: is that OK with you?

Patch-set: 5