Update patch set 2
Patch Set 2:

We have a slightly different boot and software update solution than Google, so we cannot use the recovery straight off. During the investigation of ways to handle this set of patches we concluded that it would be much less of a change to "imitate" the google master reset with a new tool - simply looking for the command file and executing the master reset - or simply return to the init flow - right after the file system is mounted.
A side effect of this is that we now do all master reset functionality with a binary measured in kilobytes instead of the 14-15mb that recovery + NAND bad block would take. 
 
Another great feature with exec is that we can halt the boot and enter "charging mode" to do charging without starting modem/wifi/ui.
 
Neither of these could be done as services.

Patch-set: 2
Reviewer: bjorn.andersson@sonyericsson.com <1003542@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0