Update patch set 3
Patch Set 3:

There are no logs from 'Deckard Autoverifier' posted to this Review - how can we see which manifest branch is used to test this patch ?
Taking the latest 
~/bin/repo init -u git://android.git.kernel.org/platform/manifest.git
Cherry picking this patch under review and building 'init' works.
Please advise on how to move forward with this change.

Patch-set: 3