Update patch set 3
Patch Set 3:

Sorry, I believe deckard built this change on top of code that was already broken.  This problem was fixed a few days ago, so that should not happen again.

* This change merges and builds in the public AOSP source tree
* This change causes a merge conflict in the internal Google source tree

There seems to be a similar change made in init.c in the internal tree, but possibly not as thorough.  Reviewers, could you please advise on how this should be merged internally?

Patch-set: 3