Update patch set 2
Patch Set 2:

This isn't a core requirement for SE Android; it is just helpful in setting ownerships on selinuxfs without having to enumerate all of the nodes (which may be dynamic).  Thought it would be useful for setting sysfs ownerships as well in similar situations.  But we can live without it if it poses a problem (curious though as to whether there is a particular concern with the change or just trying to keep init as minimalist as possible).

Patch-set: 2
Label: Code-Review=0