Update patch set 3
Patch Set 3:

I'm not sure how much use this is going to be to OEMs. They all have their own bootloaders with their own storage facilities. Even if we had this, they'd use their own. JBQ, /efs is used to share factory calibration data flashed once and not this stuff that's part of the bootloader's static data. We have had bootloader helper stuff in the past and noone used it. I'm guessing this probably works ok with TI's uboot implementation, but OEMs don't use uboot.

Patch-set: 3