Update patch set 3
Patch Set 3:

so you are using some other bootloader which only you have support for. It is doubtful that anyone else would use this format as I'm not sure how much room it leaves for the factory process, etc. I'm hesitant to pull this in with no internal users or any visibility/feedback about how this will be actually used (and hacked up) and have it bitrot until someone finds a use for them. It still sounds like something you can provide in your BSP for the time being.

Patch-set: 3