Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

(1 inline comment)

Sorry to do this after all this time, but why are you using the system property to always set this, rather than altering the in-code default to achieve the same effect?  I could see a build-specific init.rc needing a special value, but not the default one.

Patch-set: 3
Label: Code-Review=-1