Update patch set 5
Patch Set 5:

most things in init.rc that touch sys stuff should just go away. A lot of it is legacy crap from the G1, and if it's still necessary, it should be done by ueventd.<hw>.rc.

Also, i don't think we should wholesale be doing chowns on the tree. Everything will be root/root. If we want something specific to be owned by non-root, I prefer for it to be whitelisted instead.

Patch-set: 5