Update patch set 6
Patch Set 6:

We have some PnP devices, which needs to start some service or set some property on insert and stop on removal. Initially we used something like, which was defined by Google, but not implemented.

device-added-/dev/media/xxxx
    stop media
    start media

As per JP, device-added-xxx is not a good way and Google and suggested use properties instead.

Patch-set: 6