Update patch set 1
Patch Set 1:

Ignoring performance for a moment, if something is marked "extern inline" the intent is to allow inlining but if the compiler is saving space it can be assured the symbol is defined in another .o file. By making these always inline there doesn't seem to be any point to also mark them extern.

If the intent with atomic-inline is make it that the symbols are defined locally, with an inline hint, and so they don't conflict with any global symbols it seems "static inline" is closer to the billing.

Following:
http://gcc.gnu.org/onlinedocs/gcc/Inline.html#Inline
the only time that this would result in code not being inlined would be when you were compiling without optimization. Perhaps the always inline hammer can come into play if we have particular examples of GCC doing the wrong thing.

Patch-set: 1