Update patch set 2
Patch Set 2: Code-Review-1

I think this is overengineering, especially as the only known problem case is for logwrapper, which seems to be primarily for development/debugging.  Isn't the fact that there are services with logwrapper in the mako init.rc an oversight/mistake as opposed to desired functionality - it fills the logs with noise.  If my proposed change for handling logwrapper is not considered acceptable, then let us just use the existing seclabel, allow entrypoint between service domains and the logwrapper executable, and be done with it.  No change to init required.

Patch-set: 2
Reviewer: Stephen Smalley <1010111@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1