Update patch set 3
Patch Set 3:

I actually moved the maximum number of frames into the backtrace library itself and increased it to 64. I did right a test to verify that the 64 frame limit is enforced (part of backtrace_test).

Since none of the callers really cared about the maximum (they just set it to some arbitrary value), it seemed to make more sense to make the library care. It also makes it possible to change the max and everybody gets the change without needing to change all of the callers. That's at least my thought process.

Patch-set: 3