Update patch set 12
Patch Set 12:

This CL is outside of my area (I'm from audio), so please take my comments with a grain of salt:

At least in audio we don't usually accept new library code without a clear planned use in mind.

Another option would be to treat this as an upstream library (that is, something that we import into Android from outside Android, and place into external/project_name).  By treating it as an external library, it would be much easier for you to create the library speculatively and not be waiting on the Android CL review process. Provided it's hosted in a decent place such as github, and has a suitable Apache 2.0 license, then Android could import it later if & when there was a need.

Just something to consider to avoid being held up.

Patch-set: 12
Reviewer: Glenn Kasten <1004395@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0