Update patch set 7
Patch Set 7:

So I guess the question is whether it is ok to make libselinux (in Android) depend on libmincrypt, compute the hash there, and expose an interface to init for getting it.

Also might be an issue with the current layout for such a dependency, as libselinux is presently external/ whereas init and libmincrypt are system/core.  Technically I think libselinux probably ought to be moved to system/core as it isn't supposed to have external dependencies per the README.

Patch-set: 7