Update patch set 1
Patch Set 1:

Nevermind, I didn't read the whole change description.

In any case, it is very dangerous to make this function return -1.  Instead, make it measure the string in a manner consistent with what utf16_to_utf8 will do so that the buffer size will be correct even if the contents of the string are a bit mangled (but they already were to begin with).

Patch-set: 1