Update patch set 2
Patch Set 2:

Thanks to Elliot's changes, we do not need the MACRO for building for device using bionic headers. 

However we when we build the same libraries for the host, we need MACRO as the host stdint.h header file has not been changed. 

I will submit a patch with value in error.h changed to numeric literal.

Patch-set: 2