Update patch set 4
Patch Set 4:

IMHO this should be only writable to the system_server, not something writable by apps.  I don't care about reading it as it is just a hash of seapp_contexts at the time of the last restorecondata command.  Couldn't you just write it to a new /data/system file?

Patch-set: 4