Update patch set 4
Patch Set 4:

@William

Moving the timestamp out (well, replacing it with 0.0) as it is present as part of the header for all logd records. This preserves precious memory space. The sequence portion of the signature will roll over after u16 audit messages. Given that the size of the in-memory pool is 256KB, this means one will never see the rollover of the signature.

One option is for post-filtration to suck the logd timestamp back into the audit(timestamp:sequence) string. Same could be stated for the comm= and pid= squashing as they are also part of the logd headers.

Another option, if the device is shipped with a larger circular buffer devoted to recordings; at that point we should consider adjusting the squashing to use 0.1 etc to ensure that the rollover of the sequence field results in uniqueness.

Patch-set: 4