Update patch set 4
Patch Set 4:

(4 comments)

There's a lot of stuff going on in this patch including some tricky string manipulation.  The only dependency between this and the auditd patches is pid_to_name().  Please split that off into another patch to make this patch more digestible, and to unblock the auditd patches.

Patch-set: 4
Label: Verified=0