Update patch set 1
Patch Set 1:

(2 comments)

My past experience is that even in the *same* object module, the weak linkages can be overridden by stronger external symbols. If I recall, there was a different behavior between linux (overridden) and windows (was not overridden when in the same module, but was if a separate module) linkers. Using a separate module in a library is what you do always. So far, within Android, I have only used weak linkages for some socket routines in liblog as I could not link liblog against libsysutils, but I needed some of its pieces.

Patch-set: 1