Update patch set 9
Patch Set 9:

(1 comment)

Thanks to Elliott, we would like to see the sdk tools for windows updated with the requested inttypes.h. This will resolve all the PRI/SCN concerns.

We still have %z/%t formats to worry about, unless the update VS libraries are incorporated (or required to be updated for use of our sdk).

Patch-set: 9
Label: Verified=0