Update patch set 1
Patch Set 1:

Complete removal of utils/Debug.h is not KISS. Unfortunately(?) static_assert() is a C++11 thingy(a scientific term) making it possibly troublesome to replace all the COMPILE_TIME_ASSERT* calls (numbering in the dozens, so NBD to erase). I checked and they are not compiled with C++11 and are typically in frameworks which have (for me) been time consuming to gain passage.

I will look into scrubbing utils/Debug.h from the files that do not use it as a first pass towards this eventual goal?

ObligatoryDisclaimer: My thankless background task was to add -Werror to system/core and to mine for 64-bit problems; cleaning turds was merely a side effect.

Patch-set: 1