Update patch set 2
Patch Set 2: Code-Review+1 Verified+1

See https://android-review.googlesource.com/#/c/96447 where the main issue of the bug is resolved. What remains in this CL is the cache optimization of the getuid() system call, and the movement of the acquisition of the timestamp up to make it snappier.

Good job Elliott!

Patch-set: 2
Label: Code-Review=+1
Label: Verified=+1