Update patch set 1
Patch Set 1:

> > (2 comments)
 > >
 > > > I agree that the show_progress = 0 is better than return -1,
 > but
 > > > I'd prefer to keep fstat instead of switching to stat.
 > >
 > > stat() is used in two other places in this file where fstat could
 > > be used.  Should those be changed to adb_fstat() (after I make
 > > adb_fstat())?
 > 
 > yes please!

Implementing adb_fstat() on windows is turning out to be non-trivial.  I can't just call the system version of fstat because the file handle we use is the wrong type.  That means calling a bunch of other API functions to fill out the stat struct.

I've got other blocking bugs to work on, so can't afford to spend more time on this one.

Patch-set: 1