Update patch set 4
Patch Set 4:

One possible advantage of this approach over the current situation is that because the defaultservice domain is defined in AOSP policy, the CTS test will check that it passes the neverallow rules.  In contrast, any OEM-created domains are not included in the current CTS test checking.

Patch-set: 4
Reviewer: Stephen Smalley <1010111@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0