Update patch set 1
Patch Set 1:

That would be the core of a patch, but there's a bit of extra work needed as well (e.g. tidy up notificationIterators which currently has an entry per vendor in the vendor table). I should have time to put together a patch which I can submit later today.

Patch-set: 1