Update patch set 1
Patch Set 1:

(2 comments)

We're also very excited to have helped out the upstream project in this case.  Thanks for looking so quickly! :-)

On our platform, the logs were disappearing mainly due to the persistent reader problem in logd, which is now gone.  We still see tiny amounts of drops on boot (sometimes).  We're still gathering data now that the logd patch is soaking in our internal baseline.  If we see anything else interesting I'll let you know.

The intent here isn't to discount the role of optimizing the platform.  We're operating under the premise that it's ok to drop log data in liblog (can't just block writers forever), but it needs to be indicated when it happens.  One missing message in a log can cause an analysis to draw a wildly different conclusion from what actually happened.  As an OEM, we're mainly interested in log messages that are generated by upstream code, so putting retries at the app level just isn't practical and doesn't scale for configuration management reasons.

I will look into your suggestions as I find the time; hopefully soon.  Please don't let me stop you from taking it from here if you're motivated and have the time yourself.  Please do add me to any reviews you may post.

FYI, one of my colleagues expanded on the concept of the drops and introduced a periodic aggregate report of drops, which may be more useful from a platform monitoring perspective.

Patch-set: 1