Update patch set 2
Patch Set 2:

Results for BM_log_overhead_pc were caused by screen off and the 1000us sleep causing the CPU clock to be greatly reduced. Similar sleep added to the sprintf resulted in 25us rather than 1.7us. It is indicative of what the performance will be like on a sound media application that lightly loads the CPU.

Patch-set: 2