Update patch set 2
Patch Set 2:

This seems like a workaround to the fact that we have always refused to have init exec a process and then wait for it to finish, because hanging init would be catastrophic.  If we're going to allow vendor code to be linked into init we might as well implement exec instead and at least keep it out of init's memory.

Patch-set: 2