Update patch set 6
Patch Set 6: Code-Review+1 Verified+1

Thanks for all that listened to me whine. We managed to get to a point where I am only have 3x more latency than my ultimate goal; but we have at least a 4 fold or more improvement in latency for the native applications that want it and we do not violate any significant policy or cpu utilization issues in compromise.

This last change has a mix of polling for 100ms at 800us rate, followed by sem_wait sleep. I know this is gaming the benchmark. I expect that we will need to tune these for real-world users.

Patch-set: 6
Label: Code-Review=+1
Label: Verified=+1