Update patch set 1
Patch Set 1:

this is a bit odd. my arm64 clang build was fine. this only broke the x86-64 clang build.

(it's also a bit of a questionable warning anyway --- it's strictly true in that there is indeed no "return" in that function, but it's not useful because the function is "while (true) <code with no break>".)

Patch-set: 1