Update patch set 3
Patch Set 3:

Looks good, have not had an opportunity to verify.
Just some thoughts. It is a bit redundant for seclabel to be mandatory. In case where the binary is on /system and can be properly labeled it would be better to derive the context from that. If it is mandatory we risk bugs where due to inconsistencies in init vs the label on the file. Could we add --seclabel, --gid, --uid options? or what do you think nnk?

Altneratively we could use exec to start existing oneshot/disabled services.

exec mkdir /data/nnn

...
service mkdir <dynamic args to service_start>
user
group
oneshot
disabled
...

Patch-set: 3
Label: Verified=0