Update patch set 3
Patch Set 3:

I agree that seclabel should be optional, not mandatory; whenever possible, we define the transition through policy.  We only need seclabel for executables in the rootfs or for shell commands (and even the latter can be addressed if it is moved into a script with its own file security context).
It also seems that this same issue already exists for execonce, not sure when that was added.  Current SELinux policy does not allow init to execute _anything_ without changing domains, so execonce cannot work for executables in the rootfs presently if enforcing.

Patch-set: 3
Label: Verified=0