Update patch set 4
Patch Set 4: Code-Review+2 Verified+1

Found out the ./${TARBALL} was unnecessary to bootchart, dropped the ./ for KISS.

I agree, someone can always run the gnome-open or bootchart commands or equivalents again if they need to (eg: the collection is on a MAC, or a headless server?). The script is for the most common case and an easy-edit for those that are special case.

Patch-set: 4
Label: Code-Review=+2
Label: Verified=+1