Update patch set 15
Patch Set 15: Code-Review-1

This will not progress until we have document how we want the statistics to look like, discarding the statistics that have been shown to be meaningless such that we can reduce the number of lists (and thus memory used) we are collecting. Rework statistics to use a hashed list for each category to scale down further on memory usage.

Basic hash lists we expect to see once we reduce (may change if we find some other table characteristics)
- pid,uid -> "<name>" ("unknown" default for name)
-- "<name>" -> entries,size,lifetimetotalsize
- uid -> entries,size,lifetimetotalsize (for print statistics _and_ worst-uid pruning support)
- logid[index, not hashed] -> entries,size,lifetimetotalsize

Patch-set: 15
Label: Code-Review=-1