Update patch set 6
Patch Set 6: Code-Review-2

The "waiting for ueventd to process a module load firmware event" is exactly what I mentioned in our Feb 6 email. The design doc still does not address the fact that the kernel can not rely on userspace to do its bidding. I think dm-verity needs to be passed a block number to which it will write out the condition (that block could initially allocated by userspace via a file for which the fibmap is passed to dm-verity).

Patch-set: 6
Label: Code-Review=-2