Update patch set 2
Patch Set 2: Verified-1

So, submitting this is blocked on making libbase work on Windows. It's just file.cpp that doesn't compile right now, since it doesn't no DEFFILEMODE or O_CLOEXEC. We'll probably want to pull the sysdeps code out of adb and make it generally available.

Patch-set: 2
Label: Verified=-1