Update patch set 1
Patch Set 1:

I had stated in the new logger bootcamp anything that produces more than 5 message / second is viewed as a log spammer ;-/

The console spew is what is leading to the watchdog. Why do they have the console enabled? On N9 it must be _requested_ to be on, the kernel can have a default off for the messages. We do this on the N9 by turning off CONFIG_FIQ_DEBUGGER_CONSOLE_DEFAULT_ENABLE.

Patch-set: 1