Update patch set 3
Patch Set 3:

(1 comment)

> (1 comment)

Looks that the previous comment about race condition is not blocking this CL. So ps4 is ready for review.

ps4 adds another option '-a' to 'adb reboot [-a] sideload'. Without that, it reboots into sideload mode and will wait for user interactions after it finishes. If '-a' is set, it will reboot automatically after the sideload, which is for automated testing.

Patch-set: 3