Update patch set 1
Patch Set 1:

Easier said than done?

TL;DR

These files are included in support of host builds where linux/capability.h is not present or out of date. The reason is codified in include/private/android_filesystem_config.h.

Yes, we do need to get rid of this duplication. I think I stated as such; however we can not use the linux/capability.h from many of the hosts even when available, and I am _unsure_ of the damage we would do if we always include from paths bionic/libc/kernel/uapi when building for the host.

I am not sure I am comfortable littering this file with #ifdef to 'enhance' the missing content from said hosts, especially because of the mess surrounding CAP_LAST_CAP and second-guessing the structures and auto-generation that is going on here.

Patch-set: 1