Update patch set 1
Patch Set 1:

(3 comments)

I replied by email on 2014-06-24 but apparently that did not make it into the review tool's log. Repeating here, and adding inline comments.
---- 8< ----
The existing code counts an unpaired surrogate (which cannot be encoded in well-formed UTF-8) as 3 bytes UTF-8. This might work if in conversion it got replaced with U+FFFD which also takes 3 bytes.

The new code returns an error value if there is an unpaired surrogate. You would have to ensure that the callers are ready for it.

With ICU, we would call u_strToUTF8() with a NULL/capacity=0 destination buffer and just use the returned output length. That would set an error code. Or you could call a variant of this which takes a substitution character code point (u_strToUTF8WithSub()), and then its length is counted as needed.
---- >8 ----

Patch-set: 1