Update patch set 1
Patch Set 1:

(2 comments)

Sorry, but I am not completely sure what the argument is.

The change follows the pattern set in utf32_to_utf8_length, see  Unicode.h:59

The user in String8 is basically the same in the same in utf16 and utf32 and it would seem like it expects the two to be consistent.

The MediaPlayerService user (which triggered our interest) behaves well when -1 is returned.

Patch-set: 1