Update patch set 1
Patch Set 1:

gcc happily optimizes it out _without_ warning ;-} which is better?

<editorial>

The problem I have had with __attribute__((__nonnull__)) has been that it takes this optimization choice in the first place; because there are cases where one can write code that the compiler will never be able to catch a pass of a null. IMHO, this attribute should always leave any explicit checking in (with warning as per clang) unless the compiler can guarantee all paths are in-fact non-nul (eg: all calls to a static function demonstrate clearly a const char[] has been passed in)

</editorial>

Patch-set: 1