Update patch set 1
Patch Set 1:

I have just tried to set max_dgram_qlen to 20000 on emulator and it did not help: 278 lines passed one by one without a delay, then 279th was denied.
Testing art-compiler we enable many kinds of logging that produce hundreds of lines per second. The log is analyzed by QA scripts.
Neither max_dgram_qlen nor logcat -G option can help. So the solution is to print the log messages slow enough for the logging system to not lose them.

I cannot agree with your security concerns. I believe, that this kind of restrictions should be imposed solely by the logging service or socket permissions. But this patch is just a user-space feature that can be implemented in any application that uses liblog, not only in liblog itself. And we cannot stop developers from doing that. But we can implement this feature in a safe and correct way and suggest developers to use it rather than re-implement it in their own different ways.

This patch do not try to violate the limits. Instead, satisfying the limits it tries to deliver every single message.

Patch-set: 1