Update patch set 1
Patch Set 1: -Code-Review

(3 comments)

Agreed, rotation being interrupted can cause a problem. If interrupted by a system crash, it will work just fine (logcat -L notwithstanding) since the logs in memory will in fact continue where we left off since they have effectively been wiped by the reboot. If interrupted by logcat crashing (or being requested to stop! forgot about that one), it will not overwrite, it will replay the logs (a form of stutter) rather than continuing.

One of the complications could be that logcat is stopped with a -n 99, and then restarted with -n 999. This means the rotated files will have a different namespace (eg: /data/foo.01 before, and /data/foo.001 after). The only one we can count in in both namespaces is that /data/foo will always exist.

To handle all the complications, I decided it was better to just accept that sometimes there will be this stutter, in exchange for KISS.

I will prototype an algorithm that generically searches through all the files with the same basename (regardless of the namespace), finding the latest time and see if the code looks like a maintenance issue.

Patch-set: 1